requirement.txt
streamlit==1.32.0
tensorflow-cpu==2.15.0
opencv-python-headless==4.9.0.80
numpy==1.26.4
mtcnn==0.1.1
moviepy==1.0.3
librosa==0.10.1
scipy==1.11.4

